+++ src/main/java/org/interlink/grouporder/misteram/entity/MemberOrderDTO.java
@@ -3,6 +3,7 @@ package org.interlink.grouporder.misteram.entity;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -14,5 +15,5 @@ public class MemberOrderDTO {
 
     private Object items; // По идее должен быть json строкой (с)Андрей
     private int companyId;
-    private int fullPrice;
+    private BigDecimal fullPrice;
 }
